Upgrade Lucene 6.6.1 -> 7.0.1
Switch to Lucene 7.x DocValues iterator model for constructing security filter. Collapse container ID and resource ID into a single "SecurityContext" field to eliminate unnecessary (and seldom needed) lookup. Both changes should result in big security filter performance improvements.

+++ api/src/org/labkey/api/search/SearchService.java
@@ -225,7 +225,7 @@ public interface SearchService
     
     class SearchResult
     {
-        public int totalHits;
+        public long totalHits;
         public List<SearchHit> hits;
     }
